refactor: added scss styles
