+++ src/components/Product.jsx
@@ -1,4 +1,4 @@
-import "./Product.css";
+import "../styles/Product.css";
 import { Card, Elevation } from "@blueprintjs/core";
 
 const Product = ({
