+++ src/styles/Header.css
@@ -4,11 +4,11 @@
   display: flex;
   justify-content: center;
 }
-.search-input {
+.header-container .search-input {
   flex-grow: 0.9;
   border-radius: 0.5rem 0 0 0.5rem;
   font-size: 1rem;
 }
-.header-container > button {
+.header-container button {
   border-radius: 0 0.5rem 0.5rem 0;
-}
+}/*# sourceMappingURL=Header.css.map */
