+++ src/styles/Home.css
@@ -1,0 +1,6 @@
+.toggle-filter-container {
+  display: flex;
+  justify-content: space-between;
+  width: 85%;
+  margin: 1rem auto;
+}/*# sourceMappingURL=Home.css.map */
