+++ src/styles/Pagination.scss
@@ -5,7 +5,8 @@
   align-items: center;
   justify-content: center;
   margin-top: 10vh;
-}
-.pagination-container > span {
-  font-weight: 600;
+
+  span {
+    font-weight: 600;
+  }
 }
