+++ src/styles/Product.css
@@ -7,22 +7,23 @@
   overflow: hidden;
   border: 1px solid #ddd;
 }
-.product-container > img {
+.product-container img {
   width: 100%;
-  object-fit: cover;
+  -o-object-fit: cover;
+     object-fit: cover;
   height: 170px;
 }
-.product-content {
+.product-container .product-content {
   display: flex;
   position: relative;
   flex-direction: column;
   background-color: #fff;
   padding: 0.5rem;
 }
-.product-content-vendor {
+.product-container .product-content-vendor {
   font-size: 1.1rem;
 }
-.product-content-name {
+.product-container .product-content-name {
   white-space: nowrap;
   overflow: hidden;
   text-align: start;
@@ -30,30 +31,36 @@
   padding-bottom: 0.5rem;
   color: #333;
 }
-.product-content-price-container {
+.product-container .product-content-price-container {
   display: flex;
   gap: 0.5rem;
   align-items: center;
 }
-.product-content-price {
+.product-container .product-content-price {
   font-weight: 800;
 }
-.product-content-discount-price {
+.product-container .product-content-discount-price {
   color: #666;
   font-size: 0.9rem;
   text-decoration: line-through;
 }
-.product-tag-container {
+.product-container .product-tag-container {
   display: flex;
   gap: 0.5rem;
   flex-wrap: wrap;
   margin-top: 0.5rem;
 }
-.product-content-discount-percentage {
+.product-container .product-tag-container span {
+  background-color: #fff;
+  padding: 0.2rem;
+  border: 1px solid #aaa;
+  border-radius: 0.8rem;
+}
+.product-container .product-content-discount-percentage {
   color: rgb(176, 36, 36);
   font-size: 0.9rem;
 }
-.product-content-available {
+.product-container .product-content-available {
   position: absolute;
   top: 0;
   transform: translateY(-120%);
@@ -66,12 +73,6 @@
   background-color: rgb(61, 145, 61);
   color: #fff;
 }
-.out-of-stock {
+.product-container .out-of-stock {
   background-color: rgb(206, 55, 55);
-}
-.product-tag-container > span {
-  background-color: #fff;
-  padding: 0.2rem;
-  border: 1px solid #aaa;
-  border-radius: 0.8rem;
-}
+}/*# sourceMappingURL=Product.css.map */
