+++ src/styles/Product.scss
@@ -1,0 +1,78 @@
+.product-container {
+  display: flex;
+  width: 300px;
+  height: 320px;
+  flex-direction: column;
+  border-radius: 0.2rem;
+  overflow: hidden;
+  border: 1px solid #ddd;
+
+  img {
+    width: 100%;
+    object-fit: cover;
+    height: 170px;
+  }
+  .product-content {
+    display: flex;
+    position: relative;
+    flex-direction: column;
+    background-color: #fff;
+    padding: 0.5rem;
+  }
+  .product-content-vendor {
+    font-size: 1.1rem;
+  }
+  .product-content-name {
+    white-space: nowrap;
+    overflow: hidden;
+    text-align: start;
+    text-overflow: ellipsis;
+    padding-bottom: 0.5rem;
+    color: #333;
+  }
+  .product-content-price-container {
+    display: flex;
+    gap: 0.5rem;
+    align-items: center;
+  }
+  .product-content-price {
+    font-weight: 800;
+  }
+  .product-content-discount-price {
+    color: #666;
+    font-size: 0.9rem;
+    text-decoration: line-through;
+  }
+  .product-tag-container {
+    display: flex;
+    gap: 0.5rem;
+    flex-wrap: wrap;
+    margin-top: 0.5rem;
+    span {
+      background-color: #fff;
+      padding: 0.2rem;
+      border: 1px solid #aaa;
+      border-radius: 0.8rem;
+    }
+  }
+  .product-content-discount-percentage {
+    color: rgb(176, 36, 36);
+    font-size: 0.9rem;
+  }
+  .product-content-available {
+    position: absolute;
+    top: 0;
+    transform: translateY(-120%);
+    left: 0;
+    margin-left: 0.2rem;
+    padding: 0.2rem 0.5rem;
+    border-radius: 0.3rem;
+    font-weight: 600;
+    font-size: 0.9rem;
+    background-color: rgb(61, 145, 61);
+    color: #fff;
+  }
+  .out-of-stock {
+    background-color: rgb(206, 55, 55);
+  }
+}
