+++ src/styles/ProductList.css
@@ -1,0 +1,8 @@
+.product-list-container {
+  display: grid;
+  grid-template-columns: repeat(4, 1fr);
+  gap: 1rem;
+  width: 80%;
+  margin: 0 auto;
+  padding: 2rem;
+}/*# sourceMappingURL=ProductList.css.map */
