+++ src/styles/ProductList.scss
@@ -1,8 +1,9 @@
 .product-list-container {
-  display: flex;
+  display: grid;
+  grid-template-columns: repeat(4, 1fr);
   gap: 1rem;
   width: 80%;
   margin: 0 auto;
   padding: 2rem;
-  flex-wrap: wrap;
+  //   flex-wrap: wrap;
 }
