+++ src/styles/SortOptions.css
@@ -1,0 +1,12 @@
+.sort-option-container {
+  display: flex;
+  justify-content: center;
+  gap: 1rem;
+  padding: 1rem;
+  text-align: start;
+}
+.sort-option-container section {
+  text-align: start;
+  display: flex;
+  flex-direction: column;
+}/*# sourceMappingURL=SortOptions.css.map */
