+++ src/styles/SortOptions.scss
@@ -1,0 +1,12 @@
+.sort-option-container {
+  display: flex;
+  justify-content: center;
+  gap: 1rem;
+  padding: 1rem;
+  text-align: start;
+  section {
+    text-align: start;
+    display: flex;
+    flex-direction: column;
+  }
+}
